+++ src/keybindings/increase.ts
@@ -32,8 +32,8 @@ export default (logseq: ILSPluginUser) => {
         const selected = await logseq.Editor.getSelectedBlocks();
         if (selected.length > 1) {
           for (let block of selected) {
-            const content = block.content.replace(/-?[0-9]+/, (match, p1) => {
-              return `${parseInt(match) + number}`;
+            const content = block.content.replace(/^(\s*)(-?[0-9]+)/, (_match, spaces, num) => {
+              return `${spaces}${parseInt(num) + number}`;
             });
 
             await logseq.Editor.updateBlock(block.uuid, content);
@@ -41,8 +41,8 @@ export default (logseq: ILSPluginUser) => {
         } else {
           const block = await logseq.Editor.getCurrentBlock();
           if (block) {
-            const content = block.content.replace(/-?[0-9]+/, (match, p1) => {
-              return `${parseInt(match) + number}`;
+            const content = block.content.replace(/^(\s*)(-?[0-9]+)/, (_match, spaces, num) => {
+              return `${spaces}${parseInt(num) + number}`;
             });
 
             await logseq.Editor.updateBlock(block.uuid, content);